+++ webpack.config.js
@@ -1,0 +1,69 @@
+const webpack = require('webpack');
+const path = require('path');
+const CompressionPlugin = require('compression-webpack-plugin');
+
+const config = {
+    entry: './src/app.js',
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'build.js'
+    },
+    devServer: {
+        proxy: {
+            "/api/**": "http://localhost:8080",
+        },
+        contentBase: 'dist',
+        inline: true,
+        hot: true,
+        historyApiFallback: true
+    },
+    module: {
+        loaders: [{
+                test: /\.css$/,
+                loaders: ['style-loader', 'css-loader']
+            },
+            {
+                test: /.jsx?$/,
+                loader: 'babel-loader',
+                exclude: /node_modules/,
+                query: {
+                    presets: ['es2015', 'react']
+                }
+            }
+        ]
+    },
+    plugins: []
+};
+
+if(process.env.NODE_ENV == 'production') {
+  config.plugins = [
+    new webpack.DefinePlugin({
+      'process.env': {
+        'NODE_ENV': JSON.stringify('production')
+      }
+    }),
+    new webpack.optimize.UglifyJsPlugin({
+      comments: false,
+      compress: {
+        unused: true,
+        dead_code: true,
+        warnings: false,
+        drop_debugger: true,
+        conditionals: true,
+        evaluate: true,
+        sequences: true,
+        booleans: true,
+      }
+    }),
+    new webpack.optimize.AggressiveMergingPlugin(),
+    new CompressionPlugin({
+			asset: "[path].gz[query]",
+			algorithm: "gzip",
+			test: /\.(js|html)$/,
+			threshold: 10240,
+			minRatio: 0.8
+		})
+  ]
+}
+
+module.exports = config;