+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.1 (07/17/2017)
+
+- adds cli and binary option
+- basic integration with `gh-metrics`
+
 # 0.0.0
 
 - reserve `npm` name